+++ src/main/java/io/github/str4ng3r/Delete.java
@@ -16,10 +16,10 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.github.str4ng3r;
+package io.github.str4ng3r;
 
-import com.github.str4ng3r.Tables.ACTIONSQL;
-import com.github.str4ng3r.exceptions.InvalidSqlGenerationException;
+import io.github.str4ng3r.Tables.ACTIONSQL;
+import io.github.str4ng3r.exceptions.InvalidSqlGenerationException;
 
 /**
  *
@@ -40,8 +40,7 @@ public class Delete extends QueryBuilder<Delete> {
   /**
    * This should init from
    * 
-   * @param criteria
-   * @param parameters
+   * @param tableNames
    *
    * @return same object as pipe
    */
