+++ src/main/java/io/github/str4ng3r/OrderGroupBy.java
@@ -16,30 +16,24 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.github.str4ng3r;
-
-import java.util.ArrayList;
-import java.util.List;
+package io.github.str4ng3r;
 
 /**
  *
  * @author Pablo Eduardo Martinez Solis
  */
-public class SqlParameter {
-  public String sql;
-  public List<String> paramaters = new ArrayList<>();
-  protected Pagination p = null;
+class OrderGroupBy {
+  String orderBy;
 
-  SqlParameter(String sql, List<String> paramaters) {
-    this.sql = sql;
-    this.paramaters = paramaters;
+  public void orderBy(String orderBy, boolean descending) {
+    this.orderBy = descending ? orderBy + " DESC" : orderBy;
   }
 
-  @Override
-  public String toString() {
-    String query = "{\n\tsql: \"" + this.sql + "\",\n\tparameters: " + this.paramaters;
-    if (p != null)
-      query += ",\n\tpagination: " + p;
-    return query + "\n}";
+  public void groupBy(String groupBy) {
+    this.orderBy = groupBy;
+  }
+
+  public String write() {
+    return orderBy;
   }
 }
