+++ src/main/java/io/github/str4ng3r/Tables.java
@@ -16,15 +16,15 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.github.str4ng3r;
+package io.github.str4ng3r;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import com.github.str4ng3r.Join.JOIN;
-import com.github.str4ng3r.exceptions.InvalidSqlGenerationException;
+import io.github.str4ng3r.Join.JOIN;
+import io.github.str4ng3r.exceptions.InvalidSqlGenerationException;
 
 /**
  *
