+++ src/main/java/io/github/str4ng3r/WhereHaving.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.github.str4ng3r;
+package io.github.str4ng3r;
 
 import java.util.ArrayList;
 import java.util.List;
