+++ src/main/java/io/github/str4ng3r/exceptions/InvalidCurrentPageException.java
@@ -16,24 +16,10 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.github.str4ng3r;
+package io.github.str4ng3r.exceptions;
 
-/**
- *
- * @author Pablo Eduardo Martinez Solis
- */
-class OrderGroupBy {
-  String orderBy;
-
-  public void orderBy(String orderBy, boolean descending) {
-    this.orderBy = descending ? orderBy + " DESC" : orderBy;
-  }
-
-  public void groupBy(String groupBy) {
-    this.orderBy = groupBy;
-  }
-
-  public String write() {
-    return orderBy;
+public class InvalidCurrentPageException extends Exception {
+  public InvalidCurrentPageException(String error) {
+    super(error);
   }
 }
