+++ src/test/java/io/github/str4ng3r/UpdateTest.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.github.str4ng3r;
+package io.github.str4ng3r;
 
 import static org.junit.Assert.assertEquals;
 
@@ -23,8 +23,8 @@ import java.util.HashMap;
 
 import org.junit.Test;
 
-import com.github.str4ng3r.Join.JOIN;
-import com.github.str4ng3r.exceptions.InvalidSqlGenerationException;
+import io.github.str4ng3r.Join.JOIN;
+import io.github.str4ng3r.exceptions.InvalidSqlGenerationException;
 
 public class UpdateTest {
